Merge remote-tracking branch 'origin/caf/caf-wlan/master'

* origin/caf/caf-wlan/master:
  Cafstaging Release 1.0.0.106
  wlan:qcacld: Fix for Reading the FW Descriptor Bytes
  qcacld: ini: Enable gEnableMCCAdaptiveScheduler
  qcacld: hdd: Indicate FW crash before allow suspend
  qcacld: Drop management frame with invalid length
  qcacld/hdd: Update hddstats for transmit
  qcacld: CL 931243 915904 - update FW interface debug log file
  qcacld: Let the driver advertise the maximum associated STAs in AP mode
  qcacld: Prevent suspend when SAP having MAX clients connected
  qcacld : Remove duplicate code in SAP callback
  qcacld: Fix incorrect log for TDLS Discovery attempt reaching max value
  qcacld: HDD IPA enable local IP filtering
  qcacld: Convert time stamp into milli second in WMI logging
  qcacld: Add support for TDLS Off-channel and P-UAPSD
  qcacld: Request for credits for WMI commands Sent
  qcacld: fix crash due NULL pointer access on chan

+++ CORE/SERVICES/WMI/wmi_unified_priv.h
@@ -49,12 +49,12 @@ typedef adf_nbuf_t wmi_buf_t;
 struct wmi_command_debug{
 	u_int32_t command;
 	u_int32_t data[4]; /*16 bytes of WMI cmd excluding TLV and WMI headers*/
-	unsigned long time;
+	u_int32_t time;
 };
 
 struct wmi_event_debug{
 	u_int32_t event;
-	unsigned long time;
+	u_int32_t time;
 };
 
 #endif /*WMI_INTERFACE_EVENT_LOGGING*/
